requirements.txt
# MCP Server Pessoal - Dependências v1.0.0

# Core MCP
mcp>=1.2.0
fastmcp>=0.6.0

# Configuration & Validation
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Google Calendar Integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# File System & I/O
aiofiles>=23.0.0

# HTTP Client & Utilities  
httpx>=0.24.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Security
cryptography>=41.0.0
